+++ src/components/footer/desktop-footer.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
 import Link from 'next/link';
 
-import { ModalContent } from '../newsletter/ModalContent';
+import { NewsletterModals } from '../newsletter/newsletter-modals';
 import { SocialButtons } from './social-buttons';
 
 export function DesktopFooter() {
@@ -9,7 +9,7 @@ export function DesktopFooter() {
     <footer className='hidden md:flex p-10 justify-between items-center gap-10 flex-row'>
       <SocialButtons />
       <div className='absolute left-1/2 -translate-x-1/2'>
-        <ModalContent />
+        <NewsletterModals />
       </div>
       <div className='flex items-center gap-10'>
         <Link href='https://schdesign.hu' className='brand-link' target='blank'>
